+++ src/eradiate/cli.py
@@ -16,7 +16,7 @@ from ruamel.yaml import YAML
 import eradiate
 from eradiate.exceptions import DataError
 
-console = Console()
+console = Console(color_system=None)
 
 
 @click.group()
@@ -44,28 +44,40 @@ def show():
     """
     Display information useful for debugging.
     """
-    mi.set_variant("scalar_rgb")
+    import eradiate.util.sys_info
+
+    console = Console(color_system=None)
+    sys_info = eradiate.util.sys_info.show()
 
     def section(title, newline=True):
         if newline:
             console.print()
-        console.rule(title)
+        console.rule("── " + title, align="left")
         console.print()
 
     def message(text):
         console.print(text)
 
-    section("Versions", newline=False)
+    section("System", newline=False)
+    message(f"CPU: {sys_info['cpu_info']}")
+    message(f"OS: {sys_info['os']}")
+    message(f"Python: {sys_info['python']}")
+
+    section("Versions")
     message(f"• eradiate {eradiate.__version__}")
-    message(f"• mitsuba {mi.MI_VERSION}")
+    message(f"• drjit {sys_info['drjit_version']}")
+    message(f"• mitsuba {sys_info['mitsuba_version']}")
 
     section("Available Mitsuba variants")
     message("\n".join([f"• {variant}" for variant in mi.variants()]))
 
     section("Configuration")
-    for var in [x.name for x in eradiate.config.__attrs_attrs__]:
+    for var in sorted(x.name for x in eradiate.config.__attrs_attrs__):
         value = getattr(eradiate.config, var)
-        var_repr = f"{value!r}" if var == "progress" else str(value)
+        if var == "progress":
+            var_repr = f"{str(value)} ({value.value})"
+        else:
+            var_repr = str(value)
         message(f"• ERADIATE_{var.upper()}: {var_repr}")